Simplify views + fix some bugs
I'm mostly just going through the code and tidying it up right now. This
is a commit to save my progress. I'm thinking eventually about making a
neat little diagram to show the relationships in the database and code
for the coursework.

Next step is to tidy up some more, and maybe move code to modals rather
than controllers / views.